refactor(card liquid stake) refactor creating some components to handle items, change some archives names
