+++ src/modules/vault/components/liquid-stake/MobileDropdownLiquidStake.tsx
@@ -1,0 +1,56 @@
+import {
+  Drawer,
+  DrawerBody,
+  DrawerContent,
+  DrawerHeader,
+  DrawerOverlay,
+  HStack,
+  Icon,
+  Text,
+  VStack,
+} from '@chakra-ui/react';
+
+import { FuelIcon, RigIcon } from '@/components';
+
+interface MobileItemLiquidStakeProps {
+  isOpen: boolean;
+  onClose: () => void;
+  children?: React.ReactNode;
+}
+
+export function MobileDropdownLiquidStake({
+  isOpen,
+  onClose,
+  children,
+}: MobileItemLiquidStakeProps) {
+  return (
+    <Drawer placement="bottom" onClose={onClose} isOpen={isOpen}>
+      <DrawerOverlay />
+      <DrawerContent padding={4}>
+        <DrawerHeader>
+          <HStack marginBottom={4} fontWeight="normal">
+            <Icon as={FuelIcon} fontSize={24} />
+            <VStack alignItems="flex-start" gap={0}>
+              <Text fontSize={12}>Auto Stake $FUEL</Text>
+              <Text fontSize={10} color={'gray.400'}>
+                Earn Up to 18% More than Manual Staking
+              </Text>
+            </VStack>
+            <HStack
+              flex={1}
+              justifyContent="flex-end"
+              alignItems="center"
+              display={{ base: 'flex', md: 'none' }}
+            >
+              <Text fontSize={10}>powered by</Text>
+              <RigIcon fontSize={32} />
+            </HStack>
+          </HStack>
+        </DrawerHeader>
+        <DrawerBody>
+          <VStack alignItems="flex-start">{children}</VStack>
+        </DrawerBody>
+      </DrawerContent>
+    </Drawer>
+  );
+}
