+++ src/modules/vault/components/liquid-stake/ModalWithdrawals.tsx
@@ -1,5 +1,4 @@
 import {
-  Box,
   Button,
   Drawer,
   DrawerBody,
@@ -7,13 +6,14 @@ import {
   DrawerHeader,
   DrawerOverlay,
   HStack,
-  Image,
   Text,
   VStack,
 } from '@chakra-ui/react';
 
 import { Dialog } from '@/components';
 import { useScreenSize } from '@/modules/core';
+
+import { ItemWithdrawals } from './ItemWithdrawals';
 
 const TITLE = 'Withdrawals';
 const DESCRIPTION = `Rig hasn’t enabled direct stFUEL withdrawals yet. Until that option is live, you can swap stFUEL for FUEL through MIRA.`;
@@ -34,54 +34,6 @@ const ITENS = [
     description: 'The FUEL will be credited in your vault within seconds.',
   },
 ];
-
-interface WithdrawalsLiquidStakeItemProps {
-  image: string;
-  title: string;
-  description: string;
-}
-
-const WithdrawalsLiquidStakeItem = ({
-  image,
-  title,
-  description,
-}: WithdrawalsLiquidStakeItemProps) => (
-  <Box
-    width={{ base: 'full', md: 300 }}
-    marginY={4}
-    marginX={{ base: 4, md: 0 }}
-    display="flex"
-    flexDirection={{ base: 'row', md: 'column' }}
-    justifyContent={{ base: 'flex-start', md: 'center' }}
-    alignItems={{ base: 'flex-start', md: 'center' }}
-  >
-    <Image
-      src={image}
-      maxWidth={{ base: '48px', md: '80px' }}
-      alt={title}
-      borderRadius={8}
-      marginBottom={{ base: 0, md: 12 }}
-    />
-    <Box display={{ base: 'row' }} marginLeft={{ base: 4, md: 0 }}>
-      <Text
-        fontSize={16}
-        fontWeight={700}
-        marginBottom={{ base: 0, md: 6 }}
-        align={{ base: 'left', md: 'center' }}
-      >
-        {title}
-      </Text>
-      <Text
-        fontWeight={{ base: 'normal', md: 600 }}
-        fontSize={12}
-        color={'#868079'}
-        align={{ base: 'left', md: 'center' }}
-      >
-        {description}
-      </Text>
-    </Box>
-  </Box>
-);
 
 interface ModalWithdrawalsLiquidStakeProps {
   isOpen?: boolean;
@@ -116,7 +68,7 @@ export function ModalWithdrawalsLiquidStake({
           <DrawerBody>
             <VStack>
               {ITENS.map(({ image, title, description }, i) => (
-                <WithdrawalsLiquidStakeItem
+                <ItemWithdrawals
                   key={`WithdrawalsLiquidStakeItem-${i}`}
                   image={image}
                   title={title}
@@ -124,7 +76,11 @@ export function ModalWithdrawalsLiquidStake({
                 />
               ))}
             </VStack>
-            <Button variant="primary" width="full">
+            <Button
+              variant="primary"
+              width="full"
+              onClick={() => (window.location.href = 'https://mira.ly')}
+            >
               Open MIRA
             </Button>
           </DrawerBody>
@@ -155,7 +111,7 @@ export function ModalWithdrawalsLiquidStake({
         />
         <HStack marginY={{ base: 10 }}>
           {ITENS.map(({ image, title, description }, i) => (
-            <WithdrawalsLiquidStakeItem
+            <ItemWithdrawals
               key={`WithdrawalsLiquidStakeItem-${i}`}
               image={image}
               title={title}
@@ -164,7 +120,11 @@ export function ModalWithdrawalsLiquidStake({
           ))}
         </HStack>
         <Dialog.Actions hideDivider={true}>
-          <Button variant="primary" width="full">
+          <Button
+            variant="primary"
+            width="full"
+            onClick={() => window.open('https://mira.ly', '_blank')}
+          >
             Open MIRA
           </Button>
         </Dialog.Actions>
